requirements.txt
# Requires Python 3.10.9
python-telegram-bot==22.0
pymongo==4.12.0
